Don't wrap struct padding in __bindgen_marker_Opaque
Just as with zero-sized structs, this usage of blob() doesn't mean the
type is opaque: padding is perfectly safe to allocate, modify, and pass
around. By marking it with __bindgen_marker_Opaque, we prevent autocxx
from making padded structs POD.

This hasn't cropped up so far because explicit padding fields are quite
rare in bindgen, but they do show up when a struct ends with a bitfield.
I'm about to contribute POD bitfield support to autocxx, which this fix
helps enable.